Leverage ActiveRecord's transaction block to manage savepoints
Fix issue closing and opening new transaction block in BufferedCommitter...
Savepoint was created, command (insert, delete, etc.) was issued (and called "commit")
If the change_counter was at the commit_frequency, the existing transaction would be committed and a new one would be created.
The first command in the new transaction was to release or rollback the previous savepoint which would raise an exception.

https://bugzilla.redhat.com/show_bug.cgi?id=1119453